executor: speedup unit tests by spliting them into separate suites
