+++ executor/ddl_test.go
@@ -34,7 +34,7 @@ import (
 	"github.com/pingcap/tidb/util/testkit"
 )
 
-func (s *testSuite) TestTruncateTable(c *C) {
+func (s *testSuite3) TestTruncateTable(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec(`drop table if exists truncate_test;`)
@@ -51,7 +51,7 @@ func (s *testSuite) TestTruncateTable(c *C) {
 //  1. Execute the SQL of "begin";
 //  2. A SQL that will fail to execute;
 //  3. Execute DDL.
-func (s *testSuite) TestInTxnExecDDLFail(c *C) {
+func (s *testSuite3) TestInTxnExecDDLFail(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table t (i int key);")
@@ -64,7 +64,7 @@ func (s *testSuite) TestInTxnExecDDLFail(c *C) {
 	result.Check(testkit.Rows("1"))
 }
 
-func (s *testSuite) TestCreateTable(c *C) {
+func (s *testSuite3) TestCreateTable(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	// Test create an exist database
@@ -135,7 +135,7 @@ func (s *testSuite) TestCreateTable(c *C) {
 	r.Check(testkit.Rows("1000 aa"))
 }
 
-func (s *testSuite) TestCreateView(c *C) {
+func (s *testSuite3) TestCreateView(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	//create an source table
@@ -183,7 +183,7 @@ func (s *testSuite) TestCreateView(c *C) {
 	c.Assert(err.Error(), Equals, ddl.ErrViewWrongList.Error())
 }
 
-func (s *testSuite) TestCreateDropDatabase(c *C) {
+func (s *testSuite3) TestCreateDropDatabase(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("create database if not exists drop_test;")
 	tk.MustExec("drop database if exists drop_test;")
@@ -192,14 +192,14 @@ func (s *testSuite) TestCreateDropDatabase(c *C) {
 	tk.MustExec("drop database drop_test;")
 	_, err := tk.Exec("drop table t;")
 	c.Assert(err.Error(), Equals, plannercore.ErrNoDB.Error())
-	_, err = tk.Exec("select * from t;")
+	err = tk.ExecNoRes("select * from t;")
 	c.Assert(err.Error(), Equals, plannercore.ErrNoDB.Error())
 
 	_, err = tk.Exec("drop database mysql")
 	c.Assert(err, NotNil)
 }
 
-func (s *testSuite) TestCreateDropTable(c *C) {
+func (s *testSuite3) TestCreateDropTable(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table if not exists drop_test (a int)")
@@ -211,7 +211,7 @@ func (s *testSuite) TestCreateDropTable(c *C) {
 	c.Assert(err, NotNil)
 }
 
-func (s *testSuite) TestCreateDropIndex(c *C) {
+func (s *testSuite3) TestCreateDropIndex(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table if not exists drop_test (a int)")
@@ -220,14 +220,14 @@ func (s *testSuite) TestCreateDropIndex(c *C) {
 	tk.MustExec("drop table drop_test")
 }
 
-func (s *testSuite) TestAlterTableAddColumn(c *C) {
+func (s *testSuite3) TestAlterTableAddColumn(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table if not exists alter_test (c1 int)")
 	tk.MustExec("insert into alter_test values(1)")
 	tk.MustExec("alter table alter_test add column c2 timestamp default current_timestamp")
-	time.Sleep(1 * time.Second)
-	now := time.Now().Add(-time.Duration(1 * time.Second)).Format(types.TimeFormat)
+	time.Sleep(1 * time.Millisecond)
+	now := time.Now().Add(-time.Duration(1 * time.Millisecond)).Format(types.TimeFormat)
 	r, err := tk.Exec("select c2 from alter_test")
 	c.Assert(err, IsNil)
 	chk := r.NewChunk()
@@ -236,11 +236,12 @@ func (s *testSuite) TestAlterTableAddColumn(c *C) {
 	row := chk.GetRow(0)
 	c.Assert(row.Len(), Equals, 1)
 	c.Assert(now, GreaterEqual, row.GetTime(0).String())
+	r.Close()
 	tk.MustExec("alter table alter_test add column c3 varchar(50) default 'CURRENT_TIMESTAMP'")
 	tk.MustQuery("select c3 from alter_test").Check(testkit.Rows("CURRENT_TIMESTAMP"))
 }
 
-func (s *testSuite) TestAddNotNullColumnNoDefault(c *C) {
+func (s *testSuite3) TestAddNotNullColumnNoDefault(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table nn (c1 int)")
@@ -261,7 +262,7 @@ func (s *testSuite) TestAddNotNullColumnNoDefault(c *C) {
 	tk.MustQuery("select * from nn").Check(testkit.Rows("1 0", "2 0", "3 0"))
 }
 
-func (s *testSuite) TestAlterTableModifyColumn(c *C) {
+func (s *testSuite3) TestAlterTableModifyColumn(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("drop table if exists mc")
@@ -284,7 +285,7 @@ func (s *testSuite) TestAlterTableModifyColumn(c *C) {
 	c.Assert(createSQL, Equals, expected)
 }
 
-func (s *testSuite) TestDefaultDBAfterDropCurDB(c *C) {
+func (s *testSuite3) TestDefaultDBAfterDropCurDB(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	testSQL := `create database if not exists test_db CHARACTER SET latin1 COLLATE latin1_swedish_ci;`
@@ -303,7 +304,7 @@ func (s *testSuite) TestDefaultDBAfterDropCurDB(c *C) {
 	tk.MustQuery(`select @@collation_database;`).Check(testkit.Rows("utf8_unicode_ci"))
 }
 
-func (s *testSuite) TestRenameTable(c *C) {
+func (s *testSuite3) TestRenameTable(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	tk.MustExec("create database rename1")
@@ -359,7 +360,7 @@ func (s *testSuite) TestRenameTable(c *C) {
 	tk.MustExec("drop database rename2")
 }
 
-func (s *testSuite) TestUnsupportedCharset(c *C) {
+func (s *testSuite3) TestUnsupportedCharset(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	dbName := "unsupported_charset"
 	tk.MustExec("create database " + dbName)
@@ -387,7 +388,7 @@ func (s *testSuite) TestUnsupportedCharset(c *C) {
 	tk.MustExec("drop database " + dbName)
 }
 
-func (s *testSuite) TestTooLargeIdentifierLength(c *C) {
+func (s *testSuite3) TestTooLargeIdentifierLength(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	// for database.
@@ -422,7 +423,7 @@ func (s *testSuite) TestTooLargeIdentifierLength(c *C) {
 	c.Assert(err.Error(), Equals, fmt.Sprintf("[ddl:1059]Identifier name '%s' is too long", indexName2))
 }
 
-func (s *testSuite) TestShardRowIDBits(c *C) {
+func (s *testSuite3) TestShardRowIDBits(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	tk.MustExec("use test")
@@ -457,7 +458,7 @@ func (s *testSuite) TestShardRowIDBits(c *C) {
 	tk.MustExec("alter table auto shard_row_id_bits = 0")
 }
 
-func (s *testSuite) TestMaxHandleAddIndex(c *C) {
+func (s *testSuite3) TestMaxHandleAddIndex(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	tk.MustExec("use test")
@@ -474,7 +475,7 @@ func (s *testSuite) TestMaxHandleAddIndex(c *C) {
 	tk.MustExec("admin check table t1")
 }
 
-func (s *testSuite) TestSetDDLReorgWorkerCnt(c *C) {
+func (s *testSuite3) TestSetDDLReorgWorkerCnt(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	c.Assert(variable.GetDDLReorgWorkerCounter(), Equals, int32(variable.DefTiDBDDLReorgWorkerCount))
@@ -500,7 +501,7 @@ func (s *testSuite) TestSetDDLReorgWorkerCnt(c *C) {
 	res.Check(testkit.Rows("100"))
 }
 
-func (s *testSuite) TestSetDDLReorgBatchSize(c *C) {
+func (s *testSuite3) TestSetDDLReorgBatchSize(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	c.Assert(variable.GetDDLReorgBatchSize(), Equals, int32(variable.DefTiDBDDLReorgBatchSize))
