+++ executor/grant_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/pingcap/tidb/util/testkit"
 )
 
-func (s *testSuite) TestGrantGlobal(c *C) {
+func (s *testSuite3) TestGrantGlobal(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Create a new user.
 	createUserSQL := `CREATE USER 'testGlobal'@'localhost' IDENTIFIED BY '123';`
@@ -55,7 +55,7 @@ func (s *testSuite) TestGrantGlobal(c *C) {
 	}
 }
 
-func (s *testSuite) TestGrantDBScope(c *C) {
+func (s *testSuite3) TestGrantDBScope(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Create a new user.
 	createUserSQL := `CREATE USER 'testDB'@'localhost' IDENTIFIED BY '123';`
@@ -84,7 +84,7 @@ func (s *testSuite) TestGrantDBScope(c *C) {
 	}
 }
 
-func (s *testSuite) TestWithGrantOption(c *C) {
+func (s *testSuite3) TestWithGrantOption(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Create a new user.
 	createUserSQL := `CREATE USER 'testWithGrant'@'localhost' IDENTIFIED BY '123';`
@@ -98,7 +98,7 @@ func (s *testSuite) TestWithGrantOption(c *C) {
 	tk.MustQuery("SELECT grant_priv FROM mysql.DB WHERE User=\"testWithGrant\" and host=\"localhost\" and db=\"test\"").Check(testkit.Rows("Y"))
 }
 
-func (s *testSuite) TestTableScope(c *C) {
+func (s *testSuite3) TestTableScope(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Create a new user.
 	createUserSQL := `CREATE USER 'testTbl'@'localhost' IDENTIFIED BY '123';`
@@ -136,7 +136,7 @@ func (s *testSuite) TestTableScope(c *C) {
 	}
 }
 
-func (s *testSuite) TestColumnScope(c *C) {
+func (s *testSuite3) TestColumnScope(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Create a new user.
 	createUserSQL := `CREATE USER 'testCol'@'localhost' IDENTIFIED BY '123';`
@@ -176,14 +176,14 @@ func (s *testSuite) TestColumnScope(c *C) {
 	}
 }
 
-func (s *testSuite) TestIssue2456(c *C) {
+func (s *testSuite3) TestIssue2456(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("CREATE USER 'dduser'@'%' IDENTIFIED by '123456';")
 	tk.MustExec("GRANT ALL PRIVILEGES ON `dddb_%`.* TO 'dduser'@'%';")
 	tk.MustExec("GRANT ALL PRIVILEGES ON `dddb_%`.`te%` to 'dduser'@'%';")
 }
 
-func (s *testSuite) TestNoAutoCreateUser(c *C) {
+func (s *testSuite3) TestNoAutoCreateUser(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec(`DROP USER IF EXISTS 'test'@'%'`)
 	tk.MustExec(`SET sql_mode='NO_AUTO_CREATE_USER'`)
@@ -192,7 +192,7 @@ func (s *testSuite) TestNoAutoCreateUser(c *C) {
 	c.Assert(terror.ErrorEqual(err, executor.ErrPasswordNoMatch), IsTrue)
 }
 
-func (s *testSuite) TestCreateUserWhenGrant(c *C) {
+func (s *testSuite3) TestCreateUserWhenGrant(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec(`DROP USER IF EXISTS 'test'@'%'`)
 	// This only applies to sql_mode:NO_AUTO_CREATE_USER off
@@ -205,7 +205,7 @@ func (s *testSuite) TestCreateUserWhenGrant(c *C) {
 	tk.MustExec(`DROP USER IF EXISTS 'test'@'%'`)
 }
 
-func (s *testSuite) TestIssue2654(c *C) {
+func (s *testSuite3) TestIssue2654(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec(`DROP USER IF EXISTS 'test'@'%'`)
 	tk.MustExec(`CREATE USER 'test'@'%' IDENTIFIED BY 'test'`)
@@ -214,7 +214,7 @@ func (s *testSuite) TestIssue2654(c *C) {
 	rows.Check(testkit.Rows(`test %`))
 }
 
-func (s *testSuite) TestGrantUnderANSIQuotes(c *C) {
+func (s *testSuite3) TestGrantUnderANSIQuotes(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Fix a bug that the GrantExec fails in ANSI_QUOTES sql mode
 	// The bug is caused by the improper usage of double quotes like:
