+++ executor/index_lookup_join_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/pingcap/tidb/util/testkit"
 )
 
-func (s *testSuite) TestIndexLookupJoinHang(c *C) {
+func (s *testSuite1) TestIndexLookupJoinHang(c *C) {
 	tk := testkit.NewTestKitWithInit(c, s.store)
 	tk.MustExec("create table idxJoinOuter (a int unsigned)")
 	tk.MustExec("create table idxJoinInner (a int unsigned unique)")
@@ -38,7 +38,7 @@ func (s *testSuite) TestIndexLookupJoinHang(c *C) {
 	rs.Close()
 }
 
-func (s *testSuite) TestIndexJoinUnionScan(c *C) {
+func (s *testSuite1) TestIndexJoinUnionScan(c *C) {
 	tk := testkit.NewTestKitWithInit(c, s.store)
 	tk.MustExec("create table t1(id int primary key, a int)")
 	tk.MustExec("create table t2(id int primary key, a int, b int, key idx_a(a))")
@@ -92,7 +92,7 @@ func (s *testSuite) TestIndexJoinUnionScan(c *C) {
 	tk.MustExec("rollback")
 }
 
-func (s *testSuite) TestBatchIndexJoinUnionScan(c *C) {
+func (s *testSuite1) TestBatchIndexJoinUnionScan(c *C) {
 	tk := testkit.NewTestKitWithInit(c, s.store)
 	tk.MustExec("create table t1(id int primary key, a int)")
 	tk.MustExec("create table t2(id int primary key, a int, key idx_a(a))")
