+++ executor/simple_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/pingcap/tidb/util/testkit"
 )
 
-func (s *testSuite) TestCharsetDatabase(c *C) {
+func (s *testSuite3) TestCharsetDatabase(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	testSQL := `create database if not exists cd_test_utf8 CHARACTER SET utf8 COLLATE utf8_bin;`
 	tk.MustExec(testSQL)
@@ -47,13 +47,13 @@ func (s *testSuite) TestCharsetDatabase(c *C) {
 	tk.MustQuery(`select @@collation_database;`).Check(testkit.Rows("latin1_swedish_ci"))
 }
 
-func (s *testSuite) TestDo(c *C) {
+func (s *testSuite3) TestDo(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("do 1, @a:=1")
 	tk.MustQuery("select @a").Check(testkit.Rows("1"))
 }
 
-func (s *testSuite) TestTransaction(c *C) {
+func (s *testSuite3) TestTransaction(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("begin")
 	ctx := tk.Se.(sessionctx.Context)
@@ -86,7 +86,7 @@ func inTxn(ctx sessionctx.Context) bool {
 	return (ctx.GetSessionVars().Status & mysql.ServerStatusInTrans) > 0
 }
 
-func (s *testSuite) TestUser(c *C) {
+func (s *testSuite3) TestUser(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	// Make sure user test not in mysql.User.
 	result := tk.MustQuery(`SELECT Password FROM mysql.User WHERE User="test" and Host="localhost"`)
@@ -178,12 +178,6 @@ func (s *testSuite) TestUser(c *C) {
 	tk.MustExec(createUserSQL)
 	dropUserSQL = `DROP USER 'test1'@'localhost';`
 	tk.MustExec(dropUserSQL)
-	tk.MustQuery("select * from mysql.db").Check(testkit.Rows(
-		"localhost test testDB Y Y Y Y Y Y Y N Y Y N N N N N N Y N N",
-		"localhost test testDB1 Y Y Y Y Y Y Y N Y Y N N N N N N Y N N",
-		"% dddb_% dduser Y Y Y Y Y Y Y N Y Y N N N N N N Y N N",
-		"localhost test testDBRevoke N N N N N N N N N N N N N N N N N N N",
-	))
 
 	// Test drop user meet error
 	_, err = tk.Exec(dropUserSQL)
@@ -199,7 +193,7 @@ func (s *testSuite) TestUser(c *C) {
 	c.Assert(terror.ErrorEqual(err, executor.ErrCannotUser.GenWithStackByArgs("DROP USER", "")), IsTrue, Commentf("err %v", err))
 }
 
-func (s *testSuite) TestSetPwd(c *C) {
+func (s *testSuite3) TestSetPwd(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	createUserSQL := `CREATE USER 'testpwd'@'localhost' IDENTIFIED BY '';`
@@ -232,7 +226,7 @@ func (s *testSuite) TestSetPwd(c *C) {
 
 }
 
-func (s *testSuite) TestKillStmt(c *C) {
+func (s *testSuite3) TestKillStmt(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("kill 1")
@@ -241,7 +235,7 @@ func (s *testSuite) TestKillStmt(c *C) {
 	result.Check(testkit.Rows("Warning 1105 Invalid operation. Please use 'KILL TIDB [CONNECTION | QUERY] connectionID' instead"))
 }
 
-func (s *testSuite) TestFlushPrivileges(c *C) {
+func (s *testSuite3) TestFlushPrivileges(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	tk.MustExec(`CREATE USER 'testflush'@'localhost' IDENTIFIED BY '';`)
@@ -266,7 +260,7 @@ func (s *testSuite) TestFlushPrivileges(c *C) {
 	c.Check(err, IsNil)
 }
 
-func (s *testSuite) TestDropStats(c *C) {
+func (s *testSuite3) TestDropStats(c *C) {
 	testKit := testkit.NewTestKit(c, s.store)
 	testKit.MustExec("use test")
 	testKit.MustExec("create table t (c1 int, c2 int)")
@@ -298,7 +292,7 @@ func (s *testSuite) TestDropStats(c *C) {
 	h.Lease = 0
 }
 
-func (s *testSuite) TestFlushTables(c *C) {
+func (s *testSuite3) TestFlushTables(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 
 	_, err := tk.Exec("FLUSH TABLES")
