+++ executor/statement_context_test.go
@@ -28,7 +28,7 @@ const (
 	nonStrictModeSQL = "set sql_mode = ''"
 )
 
-func (s *testSuite) TestStatementContext(c *C) {
+func (s *testSuite1) TestStatementContext(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
 	tk.MustExec("create table sc (a int)")
